+++ intelmq/lib/bot.py
@@ -145,7 +145,7 @@ class Bot(object):
                 self.process()
                 self.__error_retries_counter = 0  # reset counter
 
-                if self.parameters.rate_limit:
+                if self.parameters.rate_limit and self.type != 'oneshot':
                     self.__sleep()
 
             except exceptions.PipelineError as exc:
@@ -212,6 +212,10 @@ class Bot(object):
                         # error_procedure: pass
                         else:
                             self.__error_retries_counter = 0  # reset counter
+                # no errors, check for oneshot
+                elif self.type == 'oneshot':
+                    self.logger.info('Shutting down oneshot bot.')
+                    self.stop()
             self.__handle_sighup()
 
     def __sleep(self):
@@ -400,6 +404,7 @@ class Bot(object):
 
         if self.__bot_id in list(config.keys()):
             params = config[self.__bot_id]
+            self.type = params.get('type', 'simple')
             if 'parameters' in params:
                 params = params['parameters']
             else: